refactor: extract mod review logic
